+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog
 2.13 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added possibility to define a `header_help` message that will be displayed
+  when hovering header title.
+  [gbastien]
 
 2.12 (2020-10-02)
 -----------------